+++ src/ezdxf/math/_construct.py
@@ -302,7 +302,7 @@ def gps_to_world_mercator(longitude: float, latitude: float) -> tuple[float, flo
         longitude: represents the longitude value (East-West) in decimal degrees 
         latitude: represents the latitude value (North-South) in decimal degrees.
     
-    .. versionadded:: 1.2.1
+    .. versionadded:: 1.3.0
 
     """
     # From: https://epsg.io/4326
@@ -334,7 +334,7 @@ def world_mercator_to_gps(x: float, y: float, tol: float = 1e-6) -> tuple[float,
         y: coordinate WGS84 World Mercator
         tol: accuracy for latitude calculation
 
-    .. versionadded:: 1.2.1
+    .. versionadded:: 1.3.0
 
     """
     # From: https://epsg.io/3395